Feature/issue 45 task to seconds (#55)
* add build script file

* modify Area City Location Consume Province to support JPA

* modify Device

* fix bug

* build package on mac

* fix utf8
dao for device

* modify user to jpa

* json convert for user

* modify Order to jpa

* Update Order to jpa

* 1) add batch update devicemanagerstatus api
2) update OrderDetail
3) fix bug

* send msg with seconds

* fix Order adapter

* fix bug

* AP02 msg support second task

* add scan code for test skip wx pay flow

* fix bug for manangerStatus Default value

* add consume type for device and consume

* fix device delete bug

* createTime for OrderInfo can be null

* fix bug for https

* test for seconds

* fix bug for update device

+++ service/server/src/main/java/com/hansun/server/service/ConsumeService.java
@@ -1,6 +1,6 @@
 package com.hansun.server.service;
 
-import com.hansun.dto.Consume;
+import com.hansun.server.dto.Consume;
 import com.hansun.server.db.DataStore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -25,11 +25,11 @@ public class ConsumeService {
         return dataStore.updateConsume(consume);
     }
 
-    public void deleteConsume(int consumeID) {
+    public void deleteConsume(short consumeID) {
         dataStore.deleteConsumeByConsumeID(consumeID);
     }
 
-    public Consume getConsume(int consumeID) {
+    public Consume getConsume(short consumeID) {
         return dataStore.queryConsume(consumeID);
     }
